pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "llm_harness"
version = "0.1.0"
description = "A flexible harness for working with multiple LLM providers"
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
keywords = ["llm", "ai", "testing"]
authors = [{ name = "Your Name", email = "your.email@example.com" }]

dependencies = ["litellm>=1.0.0", "pyyaml>=6.0", "python-dotenv>=1.0.0"]

[project.optional-dependencies]
test = ["pytest>=7.0", "pytest-asyncio>=0.21.0"]
dev = ["black>=23.0", "mypy>=1.0", "ruff>=0.1.0"]

[tool.pytest.ini_options]
asyncio_mode = "strict"
asyncio_default_fixture_loop_scope = "function"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::UserWarning",
    "ignore::pytest.PytestCollectionWarning",
]

[tool.hatch.build.targets.wheel]
packages = ["src/llm_harness"]

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.ruff]
line-length = 100
target-version = "py39"
select = ["E", "F", "B", "I", "UP"]

[tool.mypy]
python_version = "3.9"
strict = true
ignore_missing_imports = true
